refactor(py): move runtime helpers to separate file
